# 8269665: Clean-up toString() methods of some primitive wrappers