--- a/src/java.base/share/classes/java/lang/Boolean.java
+++ b/src/java.base/share/classes/java/lang/Boolean.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1994, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1994, 2021, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -205,7 +205,7 @@ public final class Boolean implements java.io.Serializable,
      * @since 1.4
      */
     public static String toString(boolean b) {
-        return b ? "true" : "false";
+        return String.valueOf(b);
     }
 
     /**
@@ -218,7 +218,7 @@ public final class Boolean implements java.io.Serializable,
      */
     @Override
     public String toString() {
-        return toString(value);
+        return String.valueOf(value);
     }
 
     /**