--- a/src/java.base/share/classes/java/lang/Character.java
+++ b/src/java.base/share/classes/java/lang/Character.java
@@ -8656,8 +8656,9 @@ class Character implements java.io.Serializable, Comparable<Character>, Constabl
      *
      * @return  a string representation of this object.
      */
+    @Override
     public String toString() {
-        return String.valueOf(value);
+        return toString(value);
     }
 
     /**